refactoring sample filter config - moving to shared config class
