+++ kroxylicious-sample/src/main/java/io/kroxylicious/sample/SampleProduceRequestFilter.java
@@ -21,14 +21,13 @@ import org.apache.kafka.common.record.RecordBatch;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.ByteBufferOutputStream;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.micrometer.core.instrument.Metrics;
 import io.micrometer.core.instrument.Timer;
 
-import io.kroxylicious.proxy.config.BaseConfig;
 import io.kroxylicious.proxy.filter.KrpcFilterContext;
 import io.kroxylicious.proxy.filter.ProduceRequestFilter;
+import io.kroxylicious.sample.config.SampleFilterConfig;
 
 /**
  * A sample ProduceRequestFilter implementation, intended to demonstrate how custom filters work with
@@ -46,43 +45,11 @@ import io.kroxylicious.proxy.filter.ProduceRequestFilter;
  */
 public class SampleProduceRequestFilter implements ProduceRequestFilter {
 
-    /**
-     * The Jackson configuration object for SampleFetchResponseFilter.
-     */
-    public static class SampleProduceRequestConfig extends BaseConfig {
-
-        private final String from;
-        private final String to;
-
-        /**
-         * @param from the value to be replaced
-         * @param to the replacement value
-         */
-        public SampleProduceRequestConfig(@JsonProperty(required = true) String from, @JsonProperty(required = true) String to) {
-            this.from = from;
-            this.to = to;
-        }
-
-        /**
-         * @return the configured value to be replaced
-         */
-        public String getFrom() {
-            return from;
-        }
-
-        /**
-         * @return the configured replacement value
-         */
-        public String getTo() {
-            return to;
-        }
-    }
-
     private final String from;
     private final String to;
     private final Timer timer;
 
-    public SampleProduceRequestFilter(SampleProduceRequestConfig config) {
+    public SampleProduceRequestFilter(SampleFilterConfig config) {
         this.from = config.getFrom();
         this.to = config.getTo();
         this.timer = Timer
